+++ ts/object/src/lib/decorators/build.ts
@@ -5,8 +5,9 @@ export function build( method?:string|Object, ...args ):any {
 
     function build( target:any, name: string,  descriptor?: TypedPropertyDescriptor<Function>) {
         if ( descriptor ) throw new Error("Cannot use the build decorator on a method") 
-        if ( method === undefined ) method = `_build_${name}`
-        meta(target).property(name).default( o => o[<string>method].call(o,o) )
+        meta(target).property(name).build( method )
+
+        // meta(target).property(name).build( method )
     }
 
     if ( method instanceof Object ) {