+++ ts/object/CHANGES.md
@@ -2,7 +2,12 @@
 1.3.0 Aug 29, 2021
 - traits can now over-ride the constructor of the consuming class by decorating 
 the class using the Δdecorate property
+<<<<<<< HEAD
 - added unveil function
+=======
+- added Buildable trait
+- added `does` method to the ObjectDescriptor
+>>>>>>> AGX-14 added performBuild to PropertyDescriptor
 
 1.2.1 June 26, 2021
 - inflate will now accept a customer serializer in place of class (original behavior still funcations as before)